+++ src/managers/terrain_manager.cpp
@@ -952,7 +952,7 @@ void TerrainManager::playPutBlockSound(const Blocks& blockType) {
     SfxBlockModel* blockSfxModel =
         this->t_blockManager->getDigSoundByBlockType(blockType);
     if (blockSfxModel != nullptr) {
-      const s8 ch = this->t_soundManager->getAvailableChannel();
+      const int ch = this->t_soundManager->getAvailableChannel();
       this->t_soundManager->playSfx(blockSfxModel->category,
                                     blockSfxModel->sound, ch);
     }
@@ -966,7 +966,7 @@ void TerrainManager::playDestroyBlockSound(const Blocks& blockType) {
         this->t_blockManager->getDigSoundByBlockType(blockType);
 
     if (blockSfxModel != nullptr) {
-      const s8 ch = this->t_soundManager->getAvailableChannel();
+      const int ch = this->t_soundManager->getAvailableChannel();
       this->t_soundManager->playSfx(blockSfxModel->category,
                                     blockSfxModel->sound, ch);
     }
@@ -980,7 +980,7 @@ void TerrainManager::playBreakingBlockSound(const Blocks& blockType) {
         this->t_blockManager->getDigSoundByBlockType(blockType);
 
     if (blockSfxModel != nullptr) {
-      const s8 ch = this->t_soundManager->getAvailableChannel();
+      const int ch = this->t_soundManager->getAvailableChannel();
       this->t_soundManager->playSfx(blockSfxModel->category,
                                     blockSfxModel->sound, ch);
     }
